fix: correctly select binary name from crates with a binary and library
Previously, this action may have incorrectly identified your crate name
for crates with both a binary and a library. This issue affected
crates with the libraray preceeding the binary in Cargo.toml.

This commit fixes the bug in the jq program used to identify the crate
name for crates with both a library and a binary.

Note that crates with multiple binaries will produce undefined
release behavior.