fix: rename job from Cargo test to Test
This reflects that the step may be configured to run a different
command than `cargo test`.

Technically this change may break repositories that rely on this
job name for required status checks. I have not considered job names
to be part of this package's stable API in the past but moving forward
I will.