fix: always include binary name in release asset label
This release now includes the binary name in releases containing only a
single binary.

For example, previously projects producing a single binary would create
one asset per whitelisted target, where the asset label exactly matched
the target triple:

- `aarch64-apple-darwin`
- `x86_64-apple-darwin`
- `x86_64-unknown-linux-gnu`
- `x86_64-unknown-linux-musl`

Now, this action creates releases with a single binary in the same way
it creates releases with multiple binaries: each asset is labeled with
both the binary name and the target triple, separated by a hyphen:

- `flux-capacitor-aarch64-apple-darwin`
- `flux-capacitor-x86_64-apple-darwin`
- `flux-capacitor-x86_64-unknown-linux-gnu`
- `flux-capacitor-x86_64-unknown-linux-musl`

This action is supported by **install-github-release-binary** as of
release [v2.4.0].

This is not considered a breaking change because the opinionated nature
of this repository assumes you are using install-github-release-binary
to install assets created by this workflow.

[v2.4.0]: https://github.com/EricCrosson/install-github-release-binary/releases/tag/v2.4.0