fix: do not use the same name for multiple artifact uploads
This commit fixes a regression caused by #117. Updating
actions/upload-artifact to v4 included the breaking change:

> it is no longer possible to upload to the same named Artifact multiple
> times. You must either split the uploads into multiple Artifacts with
> different names, or only upload once.

We previously hit this new edge case only when a Cargo workspace built
multiple binaries for the same target architecture.

Now, we prefix the artifact name with the binary name to ensure each
artifact name is unique.